large package refactoring + separation of spanner read/write converter
