+++ spring-cloud-gcp-data-spanner/src/main/java/org/springframework/cloud/gcp/data/spanner/core/convert/ConversionUtils.java
@@ -1,0 +1,44 @@
+/*
+ *  Copyright 2018 original author or authors.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package org.springframework.cloud.gcp.data.spanner.core.convert;
+
+import java.lang.reflect.Array;
+
+import com.google.cloud.ByteArray;
+
+
+/**
+ * @author Balint Pato
+ * @author Chengyuan Zhao
+ */
+class ConversionUtils {
+
+	static Class boxIfNeeded(Class propertyType) {
+		if (propertyType == null) {
+			return null;
+		}
+		return propertyType.isPrimitive()
+				? Array.get(Array.newInstance(propertyType, 1), 0).getClass()
+				: propertyType;
+	}
+
+	static boolean isIterableNonByteArrayType(Class propType) {
+		return Iterable.class.isAssignableFrom(propType)
+				&& !ByteArray.class.isAssignableFrom(propType);
+	}
+
+}
