+++ spring-cloud-gcp-data-spanner/src/test/java/org/springframework/cloud/gcp/data/spanner/core/SpannerMutationFactoryImplTests.java
@@ -28,9 +28,9 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+import org.springframework.cloud.gcp.data.spanner.core.convert.SpannerConverter;
 import org.springframework.cloud.gcp.data.spanner.core.mapping.SpannerColumn;
 import org.springframework.cloud.gcp.data.spanner.core.mapping.SpannerMappingContext;
-import org.springframework.cloud.gcp.data.spanner.core.mapping.SpannerObjectMapper;
 import org.springframework.cloud.gcp.data.spanner.core.mapping.SpannerTable;
 import org.springframework.data.annotation.Id;
 import org.springframework.test.context.junit4.SpringRunner;
@@ -48,14 +48,14 @@ public class SpannerMutationFactoryImplTests {
 
 	private SpannerMappingContext mappingContext;
 
-	private SpannerObjectMapper objectMapper;
+	private SpannerConverter objectMapper;
 
 	private SpannerMutationFactoryImpl spannerMutationFactory;
 
 	@Before
 	public void setUp() {
 		this.mappingContext = new SpannerMappingContext();
-		this.objectMapper = mock(SpannerObjectMapper.class);
+		this.objectMapper = mock(SpannerConverter.class);
 		this.spannerMutationFactory = new SpannerMutationFactoryImpl(this.objectMapper,
 				this.mappingContext);
 	}
